Move utility functions to common header
