--- src/utilities/string_helper.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#include <algorithm>
-#include <string>
-
-namespace NOX::StringHelper {
-
-inline std::string toLower(std::string value) {
-    const auto customToLower = [](uint8_t c) { return static_cast<uint8_t>(std::tolower(c)); };
-    std::transform(value.begin(), value.end(), value.begin(), customToLower);
-    return value;
-}
-
-} // namespace NOX::StringHelper
